(PUP-10241) Force UTF-8 encoding on XML plist data
Apple Property List (plist) XML data is always UTF-8. When executing
commands such as `/usr/bin/dscl -plist . readall /Groups`, Puppet would
consider the command output to be US-ASCII, which could ultimately break
if the output contained non-ASCII characters. Until now this didn't seem
to be an issue, but with the addition of macOS 10.15 (Catalina) this
isn't the case anymore.

The `Puppet::Util::Plist.parse_plist` method can be called with either
XML or binary data, and based on this we can force the encoding of the
XML data to UTF-8.