(PUP-10216) Eliminate node request and use last environment
Remove the node definition request and rely on the configured
environment or last used environment during the agent run.

Previously, when Puppet failed to get its node definition, it would try
to pluginsync against the agent-configured environment (usually
'production'). This is not what we want, as it first deletes all of its
plugins, and then has to re-sync them fresh from the correct environment
after it switches over.

The environment is now written in the lastrunfile, which Puppet attempts
to read if it fails to fetch the node definition. Environment values set
with `--environment` or in puppet.conf are still prioritized over the
last used environment.

Other puppet applications such as apply also write to the lastrunfile,
so in addition to writing/reading the environment, do the same for the
run mode.